Removed iframe message and changed AdsWrapper to AdsScript.
